aider: Set the desktop background color to dark blue.
# Aider chat conversation:

USER: Change the desktop color to a dark-blue.

ASSISTANT: To change the desktop background color to dark blue, I will update the `gnome-settings.yml` file to set the `primary-color` to a dark blue color value. Let's use the hex color `#00008B` for dark blue.

Here is the diff for the change:

```diff
--- vm-setup/gnome-settings.yml
+++ vm-setup/gnome-settings.yml
@@ ... @@
    - name: Set desktop background options
      shell: "gsettings set org.gnome.desktop.background {{ item }}"
      with_items:
-        - primary-color '#27285C'
+        - primary-color '#00008B'
        - show-desktop-icons false
```

This change will set the primary color of the GNOME desktop background to dark blue.