aider: Adjusted the Microsoft repository URL to use both the version number and the codename of the installed Ubuntu version.
# Aider chat conversation:

USER: the url would be a number (e.g., 22.04) while the second parameter would be a word (e.g., jammy).

ASSISTANT: To adjust the Microsoft repository URL to use both the version number and the codename of the installed Ubuntu version, we need to modify the `vm-setup/roles/add_repo/defaults/main.yml` file. We will use two Ansible facts: `ansible_distribution_version` for the version number and `ansible_distribution_release` for the codename.

Here is the diff to make the necessary changes:

```diff
--- vm-setup/roles/add_repo/defaults/main.yml
+++ vm-setup/roles/add_repo/defaults/main.yml
@@ -11,7 +11,7 @@
   microsoft:
     filename: microsoft
     key_url: https://packages.microsoft.com/keys/microsoft.asc
     repo_name: Microsoft
-    repo_url: "deb [arch=amd64] https://packages.microsoft.com/ubuntu/{{ ansible_distribution_release }}/prod {{ ansible_distribution_release }} main"
+    repo_url: "deb [arch=amd64] https://packages.microsoft.com/ubuntu/{{ ansible_distribution_version }}/prod {{ ansible_distribution_release }} main"
```

This change will ensure that the Microsoft repository URL will dynamically match both the version number and the codename of the installed version of Ubuntu.