aider: Added task to remove old APT repository before adding the new one in `install_repo.yml`.
# Aider chat conversation:

USER: The old version in sources.list.d is not being removed. How do I make ansible replace that, not add to it.

ASSISTANT: To ensure that Ansible replaces an old version of a repository in `sources.list.d` instead of adding to it, you can use the `apt_repository` module with the `state` parameter set to `absent` for the old repository before adding the new one. This will remove the old repository entry, and then you can proceed to add the new repository entry.

Here's the diff for the `vm-setup/roles/add_repo/tasks/install_repo.yml` file to include a task that removes any existing repository file before adding the new one:

```diff
--- vm-setup/roles/add_repo/tasks/install_repo.yml
+++ vm-setup/roles/add_repo/tasks/install_repo.yml
@@ -1,5 +1,14 @@
 # This file contains the tasks to install a single repository
+- name: "Remove old APT repository for {{ repo_name }}"
+  apt_repository:
+    repo: "{{ old_repo_url | default('') }}"
+    filename: "{{ filename }}"
+    state: absent
+  when: old_repo_url is defined
+  ignore_errors: yes
+
+- name: "Validate required variables for {{ repo_name }}"
   assert:
     that:
       - key_url is defined
@@ -8,6 +17,7 @@
     fail_msg: "Required variable is not defined for {{ repo_name }}."

 - name: "Install Repository Public Key for {{ repo_name }}"
+  apt_key:
     url: "{{ key_url }}"
     state: present

@@ -15,6 +25,7 @@
 - name: "Add APT repository for {{ repo_name }}"
   apt_repository:
     repo: "{{ repo_url }}"
+    filename: "{{ filename }}"
     state: present
```

In this diff, I've added a new task at the beginning of the `install_repo.yml` file that will remove an old repository if the `old_repo_url` variable is defined. This task will ignore errors to ensure that the playbook doesn't fail if the old repository doesn't exist. The `when` condition checks if the `old_repo_url` variable is defined, which should be set to the URL of the old repository that needs to be removed.

Please note that you will need to define the `old_repo_url` variable in the appropriate place, such as in the `vars` section of your playbook or role, for each repository that you want to replace.