aider: Added debug task to display downloaded git-bug binary information.
# Aider chat conversation:

USER: How do I see which git-bug file was downloaded. There is an error when running git-bug.

ASSISTANT: To help you identify which `git-bug` binary was downloaded and to provide more information in case of an error, I will add a debug task to the `git-bug.yml` playbook that outputs the downloaded file's URL and version after the download task. This will allow you to see exactly which file was downloaded.

Here is the diff for the changes to the `vm-setup/git-bug.yml` file:

```diff
--- vm-setup/git-bug.yml
+++ vm-setup/git-bug.yml
@@ ... @@
     - name: Download the latest git-bug binary
       get_url:
         url: "{{ git_bug_download_url }}"
         dest: "{{ git_bug_binary_path }}"
         mode: '0755'
       when: git_bug_version_mismatch
+      register: git_bug_download_result

+    - name: Debug downloaded git-bug binary information
+      debug:
+        msg: "Downloaded git-bug binary from {{ git_bug_download_url }} with version {{ git_bug_latest_version }}"
+      when: git_bug_download_result is changed

     - name: Ensure git-bug is executable
       file:
         path: "{{ git_bug_binary_path }}"
         mode: '0755'
         state: file
       when: git_bug_version_mismatch
```

This change adds a new task that will output the download URL and the version of the `git-bug` binary that was downloaded. It will only display this message if the download task was executed, which is controlled by the `when: git_bug_download_result is changed` condition.