aider: Update git-bug.yml playbook to filter and download the correct binary based on OS and architecture.
# Aider chat conversation:

USER: This is downloading the first git-bug (git-bug_darwin_amd64) but it needs to download git-bug_<os>_<arch> (e.g., git-bug_linux_amd64).

ASSISTANT: To ensure that the correct `git-bug` binary for the operating system and architecture is downloaded, we need to modify the `git-bug.yml` playbook to filter the assets from the GitHub release based on the OS and architecture. We will use Ansible facts to determine the correct binary to download.

Here is the diff for the changes to the `vm-setup/git-bug.yml` file:

```diff
--- vm-setup/git-bug.yml
+++ vm-setup/git-bug.yml
@@ ... @@
     - name: Set git-bug download URL and version
       set_fact:
-        git_bug_download_url: "{{ git_bug_release.json.assets[0].browser_download_url }}"
+        git_bug_download_url: "{{ git_bug_release.json.assets | selectattr('name', 'match', git_bug_binary_pattern) | map(attribute='browser_download_url') | first }}"
         git_bug_latest_version: "{{ git_bug_release.json.tag_name }}"
+      vars:
+        git_bug_binary_pattern: "git-bug_{{ ansible_system | lower }}_{{ ansible_userspace_architecture }}"

     - name: Debug downloaded git-bug binary information
       debug:
         msg: "Downloaded git-bug binary from {{ git_bug_download_url }} with version {{ git_bug_latest_version }}"
       #when: git_bug_download_result is changed
```

This change introduces a new variable `git_bug_binary_pattern` that constructs a regex pattern to match the binary name based on the system's OS and architecture using Ansible facts `ansible_system` and `ansible_userspace_architecture`. The `selectattr` filter is then used to find the correct asset from the GitHub release assets, and the `map` filter extracts the `browser_download_url` of the matched asset. The `first` filter is used to get the URL of the first matching asset.