aider: Move the installation of curl and unzip to the common packages role.
# Aider chat conversation:

USER: Move the the installation of curl and unzip to the common packages role.

ASSISTANT: To move the installation of `curl` and `unzip` from the `nerd_font` role to the `common_pkgs` role, I will make the following changes:

1. Remove the installation of `curl` and `unzip` from the `nerd_font` role's `main.yml`.
2. Add `curl` and `unzip` to the list of packages in the `common_pkgs` role's `main.yml`.

Here are the diffs for those changes:

```diff
--- vm-setup/roles/nerd_font/tasks/main.yml
+++ vm-setup/roles/nerd_font/tasks/main.yml
@@ ... @@
-- name: Install curl and unzip
-  apt:
-    state: latest
-    update_cache: yes
-    name:
-      - curl
-      - unzip
-
 - name: Create directory for Nerd Fonts
   file:
     path: "{{ nerd_font_local_dir }}"
```

```diff
--- vm-setup/roles/common_pkgs/tasks/main.yml
+++ vm-setup/roles/common_pkgs/tasks/main.yml
@@ ... @@
   apt:
     state: latest
     update_cache: yes
     name:
+      - curl
+      - unzip
       - ansible
       - ansible-lint
       - aptitude
       - build-essential
       - ca-certificates
       - curl
       - dconf-editor
       - dkms
       - dos2unix
       - elinks
       - expect
       - git
       - jq
       - "linux-headers-{{ ansible_kernel }}"
       - markdown
       - mlocate
       - pandoc
       - parallel
       - tmux
```

These changes will ensure that `curl` and `unzip` are installed as part of the common packages, which is a more appropriate location since they are utilities that may be used by other roles or tasks as well.