aider: Add conditional check to skip downloading font file if it already exists in the `nerd_font` role.
# Aider chat conversation:

USER: Won't it try to download the file even if it exists then?

ASSISTANT: You are correct. To prevent the `nerd_font` role from attempting to download the font file if it already exists, we should add a conditional check to the download task. This way, the download will only occur if the font file is not present.

Here is the diff to update the `nerd_font` role's `main.yml` file to include the conditional check:

```diff
--- vm-setup/roles/nerd_font/tasks/main.yml
+++ vm-setup/roles/nerd_font/tasks/main.yml
@@ ... @@
 - name: Download Nerd Font
+  when: not file_check.stat.exists
   get_url:
     url: "{{ nerd_font_url }}"
     dest: "{{ nerd_font_local_filename }}"
     mode: '0644'
```

By adding the `when: not file_check.stat.exists` condition to the "Download Nerd Font" task, Ansible will skip this task if the font file already exists, thus avoiding unnecessary downloads.