aider: Rename the `common_pkgs` role to `install_pkgs`.
# Aider chat conversation:

USER: Rename the common_pkgs role to install_pkgs.

ASSISTANT: To rename the `common_pkgs` role to `install_pkgs`, we need to rename the role directory and update any references to this role in the playbooks. Here are the diffs for these changes:

```diff
--- vm-setup/sweetums.yml
+++ vm-setup/sweetums.yml
@@ ... @@
   roles:
-    - role: common_pkgs
+    - role: install_pkgs
     - role: nerd_font
     - role: add_iac
     - role: git_bug
     - role: gnome_settings
```

```diff
--- vm-setup/work.yml
+++ vm-setup/work.yml
@@ ... @@
   roles:
-    - role: common_pkgs
+    - role: install_pkgs
     - role: nerd_font
     - role: add_iac
     - role: git_bug
     - role: gnome_settings
```

Please note that the actual renaming of the directory from `common_pkgs` to `install_pkgs` within the `vm-setup/roles/` directory is a file system operation and cannot be represented as a diff. You will need to perform this action manually or with a command such as `mv vm-setup/roles/common_pkgs vm-setup/roles/install_pkgs`.