aider: Added all packages from the original list to the 'common' group in the `install_pkgs` role.
# Aider chat conversation:

USER: All of the packages in the original list need to be added to the common group.

ASSISTANT: To add all the packages from the original list to the 'common' group in the `install_pkgs` role, I will update the `defaults/main.yml` file within that role. Here is the diff for the required changes:

```diff
--- vm-setup/roles/install_pkgs/defaults/main.yml
+++ vm-setup/roles/install_pkgs/defaults/main.yml
@@ ... @@
 package_groups:
   common:
+    - ansible
+    - ansible-lint
+    - aptitude
+    - build-essential
+    - ca-certificates
+    - dconf-editor
+    - dkms
+    - dos2unix
+    - elinks
+    - expect
+    - "linux-headers-{{ ansible_kernel }}"
+    - markdown
+    - mlocate
+    - pandoc
+    - parallel
+    - tmux
+    - docker-ce
+    - docker-compose
+    - docker-scan-plugin
     - curl
     - git
     - htop
     - jq
     - tree
     - unzip
     - vim
     - wget
     - zip
```

This diff adds all the packages from the original list to the 'common' group in the `install_pkgs` role's default variables file.