aider: Remove the `rescue` block to make the playbook fail and stop executing if the validation of the requested package groups fails.
# Aider chat conversation:

USER: Is it possible to make the entire play stop if a bad group is requested?

ASSISTANT: To make the entire playbook stop if a bad group is requested in the `install_pkgs` role, we need to modify the `main.yml` file within the `install_pkgs` role to remove the `rescue` block and let the `fail` task stop the execution of the playbook. Here is the diff for the required changes:

```diff
--- vm-setup/roles/install_pkgs/tasks/main.yml
+++ vm-setup/roles/install_pkgs/tasks/main.yml
@@ ... @@
   block:
     - name: Validate requested package groups
       fail:
         msg: >
           These package groups are not defined: {{ requested_groups | difference(all_groups) | join(', ') }}.
           Available package groups: {{ all_groups | join(', ') }}
       when: requested_groups | difference(all_groups) | length > 0

-    - name: Install requested package groups
-      apt:
-        state: latest
-        update_cache: yes
-        name: "{{ package_groups[item] }}"
-      loop: "{{ requested_groups }}"
-      loop_control:
-        loop_var: item
-
-  rescue:
-    - name: Capture validation failure
-      debug:
-        msg: "Validation failed for the requested package groups. Please check the package group names and try again."
+    - name: Install requested package groups
+      apt:
+        state: latest
+        update_cache: yes
+        name: "{{ package_groups[item] }}"
+      loop: "{{ requested_groups }}"
+      loop_control:
+        loop_var: item
```

By removing the `rescue` block, the playbook will now fail and stop executing if the validation of the requested package groups fails.